Use smallint instead of tinyint

+++ htdocs/cabinetmed/sql/llx_cabinetmed_diaglec.sql
@@ -19,7 +19,7 @@ CREATE TABLE llx_cabinetmed_diaglec (
   rowid             integer AUTO_INCREMENT PRIMARY KEY,
   code              varchar(8) NOT NULL,
   label             varchar(64) NOT NULL,
-  active            tinyint DEFAULT 1  NOT NULL,
+  active            smallint DEFAULT 1  NOT NULL,
   icd			    varchar(12) NULL,
   position          integer DEFAULT 10,  
   lang				varchar(12) NULL
